sched/note: Move the circle buffer code to driver
so the format code can be shared by different transport
