+++ boards/arm/lc823450/lc823450-xgevk/configs/elf/defconfig
@@ -22,6 +22,7 @@ CONFIG_DEBUG_FULLOPT=y
 CONFIG_DEBUG_SYMBOLS=y
 CONFIG_DEV_ZERO=y
 CONFIG_DISABLE_POSIX_TIMERS=y
+CONFIG_DRIVER_NOTE=y
 CONFIG_ELF=y
 CONFIG_EXAMPLES_ELF=y
 CONFIG_FS_PROCFS=y
@@ -54,7 +55,6 @@ CONFIG_SCHED_ATEXIT=y
 CONFIG_SCHED_CHILD_STATUS=y
 CONFIG_SCHED_HAVE_PARENT=y
 CONFIG_SCHED_INSTRUMENTATION=y
-CONFIG_SCHED_INSTRUMENTATION_BUFFER=y
 CONFIG_SCHED_ONEXIT=y
 CONFIG_SCHED_ONEXIT_MAX=32
 CONFIG_SCHED_STARTHOOK=y
