+++ boards/arm/lc823450/lc823450-xgevk/configs/krndis/defconfig
@@ -37,6 +37,7 @@ CONFIG_DEV_URANDOM=y
 CONFIG_DEV_ZERO=y
 CONFIG_DISABLE_POSIX_TIMERS=y
 CONFIG_DRIVERS_AUDIO=y
+CONFIG_DRIVER_NOTE=y
 CONFIG_DVFS=y
 CONFIG_ELF=y
 CONFIG_EXAMPLES_HELLO=m
@@ -154,7 +155,6 @@ CONFIG_SCHED_HAVE_PARENT=y
 CONFIG_SCHED_HPWORK=y
 CONFIG_SCHED_HPWORKPRIORITY=192
 CONFIG_SCHED_INSTRUMENTATION=y
-CONFIG_SCHED_INSTRUMENTATION_BUFFER=y
 CONFIG_SCHED_INSTRUMENTATION_PREEMPTION=y
 CONFIG_SCHED_LPWORK=y
 CONFIG_SCHED_LPWORKPRIORITY=60
