+++ boards/arm/lc823450/lc823450-xgevk/configs/posix_spawn/defconfig
@@ -21,6 +21,7 @@ CONFIG_DEBUG_FULLOPT=y
 CONFIG_DEBUG_SYMBOLS=y
 CONFIG_DEV_ZERO=y
 CONFIG_DISABLE_POSIX_TIMERS=y
+CONFIG_DRIVER_NOTE=y
 CONFIG_ELF=y
 CONFIG_EXAMPLES_POSIXSPAWN=y
 CONFIG_FS_PROCFS=y
@@ -53,7 +54,6 @@ CONFIG_SCHED_ATEXIT=y
 CONFIG_SCHED_CHILD_STATUS=y
 CONFIG_SCHED_HAVE_PARENT=y
 CONFIG_SCHED_INSTRUMENTATION=y
-CONFIG_SCHED_INSTRUMENTATION_BUFFER=y
 CONFIG_SCHED_ONEXIT=y
 CONFIG_SCHED_ONEXIT_MAX=32
 CONFIG_SCHED_STARTHOOK=y
