+++ boards/arm/lc823450/lc823450-xgevk/configs/usb/defconfig
@@ -27,6 +27,7 @@ CONFIG_DEBUG_FULLOPT=y
 CONFIG_DEBUG_SYMBOLS=y
 CONFIG_DEV_ZERO=y
 CONFIG_DISABLE_POSIX_TIMERS=y
+CONFIG_DRIVER_NOTE=y
 CONFIG_ELF=y
 CONFIG_ELF_BUFFERSIZE=512
 CONFIG_EXAMPLES_ADC=y
@@ -107,7 +108,6 @@ CONFIG_SCHED_ATEXIT=y
 CONFIG_SCHED_CHILD_STATUS=y
 CONFIG_SCHED_HAVE_PARENT=y
 CONFIG_SCHED_INSTRUMENTATION=y
-CONFIG_SCHED_INSTRUMENTATION_BUFFER=y
 CONFIG_SCHED_ONEXIT=y
 CONFIG_SCHED_ONEXIT_MAX=32
 CONFIG_SCHED_STARTHOOK=y
