+++ boards/arm/stm32f7/stm32f746-ws/configs/nsh/defconfig
@@ -31,6 +31,7 @@ CONFIG_C99_BOOL8=y
 CONFIG_CDCACM=y
 CONFIG_CDCACM_EPINTIN_HSSIZE=512
 CONFIG_DEBUG_SYMBOLS=y
+CONFIG_DRIVER_NOTE=y
 CONFIG_FAT_DMAMEMORY=y
 CONFIG_FAT_LCNAMES=y
 CONFIG_FAT_LFN=y
@@ -69,7 +70,6 @@ CONFIG_SCHED_HPWORK=y
 CONFIG_SCHED_HPWORKPRIORITY=192
 CONFIG_SCHED_HPWORKSTACKSIZE=1800
 CONFIG_SCHED_INSTRUMENTATION=y
-CONFIG_SCHED_INSTRUMENTATION_BUFFER=y
 CONFIG_SCHED_LPWORK=y
 CONFIG_SCHED_LPWORKSTACKSIZE=1800
 CONFIG_SCHED_NOTE_BUFSIZE=512
