+++ boards/xtensa/esp32/esp32-core/configs/smp/defconfig
@@ -20,6 +20,7 @@ CONFIG_BUILTIN=y
 CONFIG_DEBUG_FEATURES=y
 CONFIG_DEBUG_FULLOPT=y
 CONFIG_DEBUG_SYMBOLS=y
+CONFIG_DRIVER_NOTE=y
 CONFIG_ESP32_UART0=y
 CONFIG_EXAMPLES_HELLO=y
 CONFIG_EXPERIMENTAL=y
@@ -44,7 +45,6 @@ CONFIG_RAW_BINARY=y
 CONFIG_READLINE_CMD_HISTORY=y
 CONFIG_RR_INTERVAL=200
 CONFIG_SCHED_INSTRUMENTATION=y
-CONFIG_SCHED_INSTRUMENTATION_BUFFER=y
 CONFIG_SDCLONE_DISABLE=y
 CONFIG_SMP=y
 CONFIG_SMP_IDLETHREAD_STACKSIZE=3072
